% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_bias.R
\name{dq_bias}
\alias{dq_bias}
\title{Data Quality - Bias}
\usage{
dq_bias(rules_results, compare_group, confidence_ref)
}
\arguments{
\item{rules_results}{output from \code{dq_check} after applying the rules on the new datasets with the compare group being defined}

\item{compare_group}{How data quality is assessed between different group_ids.}

\item{confidence_ref}{the reference standard for comparison.}
}
\value{
The rules are ranked based on the bias.
 \item{rules}{list of rules}
 \item{rule_category }{Category of rules, defined during the rule mining process}
 \item{evalulation_group }{Combination of variables specified in the \code{compare_group}, level of unit of rules comparison}
 \item{confidence }{Confidence used as the reference standard}
 \item{confi_diff_sum }{Weighted sum of rules difference}
 \item{rules_rank }{Rank of rules based on the bias (weighted sum of rules difference)}
 \item{group_id }{Level of unit for rule application}
 \item{confidence_new }{Confidence of rules in the new datasets}
}
\description{
Bias indicate how the quality of a datasets is different from a reference standard. The weighted difference of rules confidence
between new dataset and reference standard are calculated and ranked for each category of rules.
}
\details{
A reference standard is selected for bias calculation.
First, confidence intervals of rules in the new datasets are calculated.
Then, the range of confidence intervals is dervied.
Finally, difference between rule confidences in the new dataset and reference standard is calcualted and weighted by (1- range).
The smaller range, the higher weight. The sum of weighted difference is calculated and ranked based on the compare_group.
}
\examples{





}
