% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_check.R
\name{dq_check}
\alias{dq_check}
\title{Data quality check using association rules}
\usage{
dq_check(data_input, rules_input, compare_group, group_var, item_id, items)
}
\arguments{
\item{data_input}{Transaction data in long format}

\item{rules_input}{The list of association rules from the \code{rules_mining, rules_clean, rules_test or rules_bootstrap} functions}

\item{compare_group}{Group indicating how data quality will be assessed}

\item{group_var}{1 or more Variables used to group rules}

\item{item_id}{Unique identifier for each transaction}

\item{items}{The variable indicating the items in a transaction}
}
\value{
A data.frame with confidence calculated in the new dataset
  \item{rules }{A list of rules being tested}
  \item{confidence }{Rules confidence from the training data}
  \item{rule_category }{Category of rules, defined during the rule mining process}
  \item{group_id }{Level of units for rules application}
  \item{n_rhs }{Number of transactions in the right hand side of rules}
  \item{n_lhs }{Number of transactions in the left hand side of rules}
  \item{n_both }{Number of transactions in both right and left hand sides of rules}
  \item{confidence_new }{confidence calculate in the new dataset}
}
\description{
Apply the developed rules to calculate the rules confidence on the new dataset.
We developed the rules for data quality assessment purposes. The changes of rules confidences at the different comparison level
can indicate the issues of consistency and completeness of data quality.
}
\examples{






}
