% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_variance.R
\name{dq_variance}
\alias{dq_variance}
\title{Data quality - Variance}
\usage{
dq_variance(rules_results = dq_d4, compare_group = "hospital_type")
}
\arguments{
\item{compare_group}{}
}
\value{
The output is similar as dq_bias except for that rules are ranked based on variance.
}
\description{
To calculate the variance, we first calculated the 95% confidence intervals of pairwise differences
between group_id with in the same compare groups. Range of confidence intervals is calculated.
Weighted difference between group_ids were calculated and sum. The rules are ranked based on the groups.
}
\examples{





}
