% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_clean.R
\name{rules_clean}
\alias{rules_clean}
\title{Remove nested rules}
\usage{
rules_clean(rules = xx, confidence_diff = 0.05)
}
\arguments{
\item{rules}{The list of association rules from the \code{rules_mining} function}

\item{confidence_diff}{The minimum difference between nested rules}
}
\value{
The output is smilar to output from \code{rules_mining} function with the nested rules being removed
  \item{LHS }{Codes in the left hand side of rules, more than 1 code can be presented in the LHS}
  \item{RHS }{Code in the right hand side of rules. Only one code can be there for easy interpretation}
}
\description{
Two rules were considered nested if they have the same RHS and the LHS of one rules is a subset of the other.
The rules's confidences are compared. Only the simple rule is kept
if the confidence difference is less than the selected threshold \code{confidence_diff}.
}
\examples{

}
