% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_clean_sequence.R
\name{rules_clean_sequence}
\alias{rules_clean_sequence}
\title{Remove nested rules using the sequence method}
\usage{
rules_clean_sequence(rules = rules_list, confidence_diff = 0.05)
}
\arguments{
\item{rules}{The list of association rules output from the \code{rules_mining} function}

\item{confidence_diff}{The minimum difference of confidence between the two nested rules}
}
\value{
The output is smilar to output from \code{rules_mining} function with the nested rules being removed. Only the following two columns are added.
  \item{LHS }{Codes in the left hand side of rules, more than 1 code can be presented in the LHS}
  \item{RHS }{Code in the right hand side of rules. Only one code can be there for easy interpretation}
  \item{lhs_list }{List of codes in the left hand side of rules}
  \item{LHS_length }{Number of codes in the left hand side of rules}
}
\description{
Two rules were considered nested if they have the same RHS while the LHS of one rule is a subset of the LHS on the other.
Confidences between two rules are compared.
if the confidence difference is less than the selected threshold: \code{confidence_diff}, only the simple rule is kept.
}
\examples{



}
