% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_statistical_test.R
\name{rules_statistical_test}
\alias{rules_statistical_test}
\title{Statistical association test on the rules}
\usage{
rules_statistical_test(rules, data_input, group_var, item_id, items,
  method_p_adjustment = "fdr")
}
\arguments{
\item{rules}{The list of association rules from the \code{rules_mining, rules_clean, rules_test or rules_bootstrap} functions}

\item{data_input}{The training data used for the rules development}

\item{group_var}{1 or more variables used to group rules}

\item{item_id}{Unique identifier for each transaction}

\item{items}{The variable indicating the items in a transaction}

\item{method_p_adjustment}{Methods used to conduct p-value adjustment. See \code{p.adjust} for the list of methods}
}
\value{
The output is in data.frame format. See functions of \code{rules_mining, rules_clean, rules_test}
 for the definition of each column.
 \item{support_rhs }{Support for the right hand side of rules}
 \item{pvalue_adjusted }{Adjusted P value for chi-square test on the association between the RHS and LHS of rules}
}
\description{
Use of chi-square test to test the statistical association between the LHS and RHS of rules.
This function used the \code{chisq.test} function and simulated p-values are calculated.
The p values are adjusted for multiple comparisons using \code{p.adjust} function. The default method is false discovery rate.
}
\examples{





}
