% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_test.R
\name{rules_test}
\alias{rules_test}
\title{Rule test on the test or validation data}
\usage{
rules_test(rules, test_data, item_id, group_var, items)
}
\arguments{
\item{rules}{The list of association rules from the \code{rules_clean} or \code{rules_mining} functions}

\item{test_data}{Test or validation dataset}

\item{item_id}{Unique identifier for each transaction}

\item{group_var}{1 or more variables used to group rules}

\item{items}{The variable indicating the items in a transaction}
}
\value{
The output is smilar to output from \code{rules_clean} function with support and confidence calculated on the test or validation dataset
  \item{support_n_test }{number of transactions with codes from both RHS and LHS for the corresponding rule category}
  \item{lhs_n_test }{Number of transactions with codes from the LHS for the corresponding rule category}
  \item{confidence_test }{confidence of rules on the test or validation datasets}
  \item{support_test }{Support of rules on on the test or validaiton datasets}
}
\description{
Apply the list of rules on the test or validation data to calculate support and confidence on the new datasets
}
\examples{



}
