% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_WHO.R
\name{map_to_WHO}
\alias{map_to_WHO}
\title{Map the ICD-10 coded data into ICD-10 WHO version}
\usage{
map_to_WHO(data_input = dad2013, colname_dx = "dx",
  chapters_excluded = c("R", "S", "T", "U", "V", "W", "X", "Y", "Z",
  icd10_who = icd10_who))
}
\arguments{
\item{data_input}{raw data in long format}

\item{colname_dx}{colnames indicate the raw ICD-10 codes}

\item{chapters_excluded}{Chapters being excluded from analysis: the first digits of ICD-10 codes}
}
\description{
Map country-specific ICD codes to ICD-1o WHO version
}
